chore: adding first mock for phone input. also separating the screens
