+++ app/screens/phone-auth-screen/phone-auth.tsx
@@ -1,10 +1,13 @@
 /* eslint-disable react-native/no-inline-styles */
+import { gql } from "@apollo/client"
+import analytics from "@react-native-firebase/analytics"
+import { RouteProp } from "@react-navigation/native"
+import { StackNavigationProp } from "@react-navigation/stack"
+import { Button, Input } from "@rneui/base"
 import * as React from "react"
 import { LegacyRef, Ref, useCallback, useEffect, useRef, useState } from "react"
 import {
   ActivityIndicator,
-  ActivityIndicatorProps,
-  Alert,
   KeyboardAvoidingView,
   Platform,
   ScrollView,
@@ -12,37 +15,21 @@ import {
   TextInput,
   View,
 } from "react-native"
-import { Button, Input } from "@rneui/base"
-import { gql } from "@apollo/client"
 import EStyleSheet from "react-native-extended-stylesheet"
-import PhoneInput from "react-native-phone-number-input"
-import analytics from "@react-native-firebase/analytics"
-import { StackNavigationProp } from "@react-navigation/stack"
-import { RouteProp } from "@react-navigation/native"
 
-import { CloseCross } from "../../components/close-cross"
+import { UserLoginMutationHookResult, useUserLoginMutation } from "@app/graphql/generated"
+import { useIsAuthed } from "@app/graphql/is-authed-context"
+import { useI18nContext } from "@app/i18n/i18n-react"
+import crashlytics from "@react-native-firebase/crashlytics"
 import { Screen } from "../../components/screen"
+import { useAppConfig } from "../../hooks"
+import type { PhoneValidationStackParamList } from "../../navigation/stack-param-lists"
 import { color } from "../../theme"
 import { palette } from "../../theme/palette"
-import { toastShow } from "../../utils/toast"
 import BiometricWrapper from "../../utils/biometricAuthentication"
 import { AuthenticationScreenPurpose } from "../../utils/enum"
-import BadgerPhone from "./badger-phone-01.svg"
-import type { PhoneValidationStackParamList } from "../../navigation/stack-param-lists"
 import { parseTimer } from "../../utils/timer"
-import { useAppConfig, useGeetestCaptcha } from "../../hooks"
-import DownArrow from "@app/assets/icons/downarrow.svg"
-import { useI18nContext } from "@app/i18n/i18n-react"
-import { logRequestAuthCode } from "@app/utils/analytics"
-import crashlytics from "@react-native-firebase/crashlytics"
-import {
-  UserLoginMutationHookResult,
-  useCaptchaRequestAuthCodeMutation,
-  useUserLoginMutation,
-} from "@app/graphql/generated"
-import { useIsAuthed } from "@app/graphql/is-authed-context"
-
-const phoneRegex = new RegExp("^\\+[0-9]+$")
+import { toastShow } from "../../utils/toast"
 
 const styles = EStyleSheet.create({
   authCodeEntryContainer: {
@@ -62,32 +49,9 @@ const styles = EStyleSheet.create({
     width: "200rem",
   },
 
-  buttonContinue: {
-    alignSelf: "center",
-    backgroundColor: color.palette.blue,
-    width: "200rem",
-    marginVertical: "15rem",
-    padding: "15rem",
-  },
-
   codeContainer: {
     alignSelf: "center",
     width: "70%",
-  },
-
-  image: {
-    alignSelf: "center",
-    marginBottom: "30rem",
-    resizeMode: "center",
-  },
-
-  phoneEntryContainer: {
-    borderColor: color.palette.darkGrey,
-    borderRadius: 5,
-    borderWidth: 1,
-    flex: 1,
-    marginHorizontal: "40rem",
-    marginVertical: "18rem",
   },
 
   sendAgainButtonRow: {
@@ -105,17 +69,8 @@ const styles = EStyleSheet.create({
     textAlign: "center",
   },
 
-  textContainer: {
-    backgroundColor: color.transparent,
-  },
-
   textDisabledSendAgain: {
     color: color.palette.midGrey,
-  },
-
-  textEntry: {
-    color: color.palette.darkGrey,
-    fontSize: "16rem",
   },
 
   timerRow: {
@@ -125,10 +80,6 @@ const styles = EStyleSheet.create({
     textAlign: "center",
   },
 })
-
-type WelcomePhoneInputScreenProps = {
-  navigation: StackNavigationProp<PhoneValidationStackParamList, "welcomePhoneInput">
-}
 
 gql`
   mutation captchaRequestAuthCode($input: CaptchaRequestAuthCodeInput!) {
@@ -149,222 +100,6 @@ gql`
     }
   }
 `
-
-export const WelcomePhoneInputScreen: React.FC<WelcomePhoneInputScreenProps> = ({
-  navigation,
-}) => {
-  const {
-    geetestError,
-    geetestValidationData,
-    loadingRegisterCaptcha,
-    registerCaptcha,
-    resetError,
-    resetValidationData,
-  } = useGeetestCaptcha()
-
-  const { LL } = useI18nContext()
-
-  const [phoneNumber, setPhoneNumber] = useState("")
-  const { appConfig } = useAppConfig()
-
-  const phoneInputRef = useRef<PhoneInput | null>(null)
-
-  const [captchaRequestAuthCode, { loading: loadingRequestPhoneCode }] =
-    useCaptchaRequestAuthCodeMutation({
-      fetchPolicy: "no-cache",
-    })
-
-  const setPhone = (newPhoneNumber: string) => {
-    setPhoneNumber(newPhoneNumber)
-  }
-
-  useEffect(() => {
-    if (phoneNumber) {
-      // This bypasses the captcha for local dev
-      // Comment it out to test captcha locally
-      if (appConfig.galoyInstance.name === "Local") {
-        navigation.navigate("welcomePhoneValidation", { phone: phoneNumber, setPhone })
-        setPhoneNumber("")
-      } else {
-        registerCaptcha()
-      }
-    }
-  }, [appConfig.galoyInstance.name, navigation, phoneNumber, registerCaptcha])
-
-  useEffect(() => {
-    if (geetestValidationData) {
-      const sendRequestAuthCode = async () => {
-        try {
-          const input = {
-            phone: phoneNumber,
-            challengeCode: geetestValidationData?.geetestChallenge,
-            validationCode: geetestValidationData?.geetestValidate,
-            secCode: geetestValidationData?.geetestSecCode,
-          }
-          resetValidationData()
-          logRequestAuthCode(appConfig.galoyInstance.name)
-
-          const { data } = await captchaRequestAuthCode({ variables: { input } })
-
-          if (!data) {
-            toastShow({
-              message: (translations) => translations.errors.generic(),
-              currentTranslation: LL,
-            })
-            return
-          }
-
-          if (data.captchaRequestAuthCode.success) {
-            navigation.navigate("welcomePhoneValidation", {
-              phone: phoneNumber,
-              setPhone,
-            })
-            setPhoneNumber("")
-          } else if ((data?.captchaRequestAuthCode?.errors?.length || 0) > 0) {
-            const errorMessage = data.captchaRequestAuthCode.errors[0].message
-            if (errorMessage === "Too many requests") {
-              toastShow({
-                message: (translations) => translations.errors.tooManyRequestsPhoneCode(),
-                currentTranslation: LL,
-              })
-            } else {
-              toastShow({ message: errorMessage })
-            }
-          } else {
-            toastShow({
-              message: (translations) => translations.errors.generic(),
-              currentTranslation: LL,
-            })
-          }
-        } catch (err: unknown) {
-          if (err instanceof Error) {
-            crashlytics().recordError(err)
-            console.debug({ err })
-            toastShow({
-              message: (translations) => translations.errors.generic(),
-              currentTranslation: LL,
-            })
-          }
-        }
-      }
-      sendRequestAuthCode()
-    }
-  }, [
-    geetestValidationData,
-    navigation,
-    phoneNumber,
-    setPhoneNumber,
-    captchaRequestAuthCode,
-    resetValidationData,
-    appConfig.galoyInstance.name,
-    LL,
-  ])
-
-  useEffect(() => {
-    if (geetestError) {
-      const error = geetestError
-      resetError()
-      toastShow({ message: error })
-    }
-  })
-
-  const submitPhoneNumber = () => {
-    if (!phoneInputRef.current) {
-      return
-    }
-
-    const phone = phoneInputRef.current.state.number
-
-    const formattedNumber = phoneInputRef.current.getNumberAfterPossiblyEliminatingZero()
-
-    const cleanFormattedNumber = formattedNumber.formattedNumber.replace(/[^\d+]/g, "")
-
-    if (
-      !phoneInputRef.current.isValidNumber(phone) ||
-      !phoneRegex.test(cleanFormattedNumber)
-    ) {
-      Alert.alert(`${phone} ${LL.errors.invalidPhoneNumber()}`)
-      return
-    }
-
-    setPhoneNumber(cleanFormattedNumber)
-  }
-
-  const showCaptcha = phoneNumber.length > 0
-  let captchaContent: ReturnType<React.FC<ActivityIndicatorProps>> | null
-
-  if (loadingRegisterCaptcha || loadingRequestPhoneCode) {
-    captchaContent = <ActivityIndicator size="large" color={color.primary} />
-  } else {
-    captchaContent = null
-  }
-
-  const renderDropdownImage = () => {
-    return <DownArrow testID="DropDownButton" width={12} height={14} />
-  }
-
-  return (
-    <Screen backgroundColor={palette.lighterGrey} preset="scroll">
-      <View style={{ flex: 1, justifyContent: "space-around", marginTop: 50 }}>
-        <View>
-          <BadgerPhone style={styles.image} />
-          <Text style={styles.text}>
-            {showCaptcha
-              ? LL.WelcomePhoneInputScreen.headerVerify()
-              : LL.WelcomePhoneInputScreen.header()}
-          </Text>
-        </View>
-        {showCaptcha ? (
-          captchaContent
-        ) : (
-          <KeyboardAvoidingView>
-            <PhoneInput
-              ref={phoneInputRef}
-              value={phoneNumber}
-              containerStyle={styles.phoneEntryContainer}
-              textInputStyle={styles.textEntry}
-              textContainerStyle={styles.textContainer}
-              defaultValue={phoneNumber}
-              defaultCode="SV"
-              layout="first"
-              renderDropdownImage={renderDropdownImage()}
-              textInputProps={{
-                placeholder: LL.WelcomePhoneInputScreen.placeholder(),
-                returnKeyType: loadingRequestPhoneCode ? "default" : "done",
-                onSubmitEditing: submitPhoneNumber,
-                keyboardType: "phone-pad",
-                textContentType: "telephoneNumber",
-                accessibilityLabel: "Input phone number",
-              }}
-              countryPickerProps={{
-                modalProps: {
-                  testID: "country-picker",
-                },
-              }}
-              codeTextStyle={{ marginLeft: -25 }}
-              autoFocus
-            />
-            <ActivityIndicator
-              animating={loadingRequestPhoneCode}
-              size="large"
-              color={color.primary}
-              style={{ marginTop: 32 }}
-            />
-          </KeyboardAvoidingView>
-        )}
-        <Button
-          buttonStyle={styles.buttonContinue}
-          title={LL.WelcomePhoneInputScreen.continue()}
-          disabled={Boolean(phoneNumber)}
-          onPress={() => {
-            submitPhoneNumber()
-          }}
-        />
-      </View>
-      <CloseCross color={palette.darkGrey} onPress={() => navigation.goBack()} />
-    </Screen>
-  )
-}
 
 type WelcomePhoneValidationScreenDataInjectedProps = {
   navigation: StackNavigationProp<PhoneValidationStackParamList, "welcomePhoneValidation">
